generalize events
