+++ internal/handlers/receiver.go
@@ -1,0 +1,30 @@
+package handlers
+
+import (
+	"context"
+	"net/http"
+
+	awsevents "github.com/aws/aws-lambda-go/events"
+	"github.com/care-giver-app/care-giver-api/internal/receiver"
+	"github.com/care-giver-app/care-giver-api/internal/response"
+	"go.uber.org/zap"
+)
+
+func HandleReceiver(ctx context.Context, params HandlerParams) (awsevents.APIGatewayProxyResponse, error) {
+	params.AppCfg.Logger.Info("handling get receiver event")
+
+	rid, err := validatePathParameters(params.Request, receiver.ParamID, receiver.DBPrefix)
+	if err != nil {
+		params.AppCfg.Logger.Error("error validating path parameters", zap.Error(err))
+		return response.CreateBadRequestResponse(), nil
+	}
+
+	r, err := params.ReceiverRepo.GetReceiver(rid)
+	if err != nil {
+		params.AppCfg.Logger.Error("error retrieving receiver from db", zap.Error(err))
+		return response.CreateInternalServerErrorResponse(), nil
+	}
+
+	params.AppCfg.Logger.Info("processed get receiver event successfully")
+	return response.FormatResponse(r, http.StatusOK), nil
+}
