+++ main.go
@@ -25,7 +25,7 @@ var (
 	appCfg          *appconfig.AppConfig
 	userRepo        *repository.UserRepository
 	receiverRepo    *repository.ReceiverRepository
-  eventRepo    *repository.EventRepository
+	eventRepo       *repository.EventRepository
 	handlerRegistry handlers.RegistryProvider
 )
 
@@ -50,9 +50,9 @@ func init() {
 
 	appCfg.Logger.Info("initializing event repository")
 	eventRepo = repository.NewEventRespository(context.TODO(), appCfg, dynamoClient)
-  
+
 	appCfg.Logger.Info("initializing handler registry")
-	handlerRegistry = handlers.NewRegistry(appCfg, userRepo, receiverRepo)
+	handlerRegistry = handlers.NewRegistry(appCfg, userRepo, receiverRepo, eventRepo)
 }
 
 func handler(ctx context.Context, req events.APIGatewayProxyRequest) (events.APIGatewayProxyResponse, error) {
